rename and organize tests folder
