+++ tests/component/LandingPage.spec.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, it, expect, vitest } from 'vitest'
 import LandingPage from '../pages/index.vue'
 import { DOMWrapper, mount } from '@vue/test-utils'
 import { createTestingPinia } from '@pinia/testing'
-import { useAuthStore } from '../store/auth'
+import { useAuthStore } from '../../store/auth'
 import flushPromises from 'flush-promises'
 import { createRouter, createMemoryHistory, RouteRecordRaw } from 'vue-router'
 import HomePage from '../pages/HomePage.vue'
