+++ vitest.config.ts
@@ -3,6 +3,6 @@ import { defineVitestConfig } from 'nuxt-vitest/config'
 export default defineVitestConfig({
   test: {
     environment: 'nuxt',
-    setupFiles: './__tests__/vue-test-utils-plugins.ts'
+    setupFiles: './tests/vue-test-utils-plugins.ts'
   },
 })
